+++ src/Data/Pool.hs
@@ -73,10 +73,10 @@ takeResource pool = mask_ $ do
         q <- newEmptyTMVar
         writeTVar (stripeVar lp) $! stripe {queueR = Queue q (queueR stripe)}
         pure
-          $ waitForResource (stripeVar lp) q >>= \case
+          $ waitForResource lp q >>= \case
             Just a -> pure (a, lp)
             Nothing -> do
-              a <- createResource (poolConfig pool) `onException` restoreSize (stripeVar lp)
+              a <- createResource (poolConfig pool) `onException` restoreSize lp
               pure (a, lp)
       else takeAvailableResource pool lp stripe
 
@@ -131,7 +131,7 @@ takeAvailableResource pool lp stripe = case cache stripe of
   [] -> do
     writeTVar (stripeVar lp) $! stripe {available = available stripe - 1}
     pure $ do
-      a <- createResource (poolConfig pool) `onException` restoreSize (stripeVar lp)
+      a <- createResource (poolConfig pool) `onException` restoreSize lp
       pure (a, lp)
   Entry a _ : as -> do
     writeTVar (stripeVar lp)